Time: 0 ms (100%), Space: 25.9 MB (6.48%) - LeetHub

+++ 3194-minimum-average-of-smallest-and-largest-elements/3194-minimum-average-of-smallest-and-largest-elements.cpp
@@ -3,14 +3,14 @@ public:
     double minimumAverage(vector<int>& nums) {
         
         int n= nums.size();
-        vector<float> answers;
+        vector<long double> answers;
 
         sort(nums.begin(), nums.end());
         int left =0, right=n-1;
-        double min_avg = DBL_MAX;
+        long double min_avg = DBL_MAX;
 
         while(left < right){
-            double avg = (nums[left] + nums[right])/2.0;
+            long double avg = (nums[left] + nums[right])/2.0;
             answers.push_back(avg);
             min_avg = min(min_avg, avg);
             left++;
